pyproject.toml
[tool.poetry]
name = "valence_api"
version = "0.1.0"
description = ""
authors = ["RasterMk2 <70229194+RasterMk2@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.79.0"}
tortoise-orm = {extras = ["accel", "sqlite"], version = "^0.19.2"}
typer = {extras = ["all"], version = "^0.6.1"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
types-python-jose = "^3.3.4"
passlib = {extras = ["argon2"], version = "^1.7.4"}
types-passlib = "^1.7.7"
python-dotenv = "^0.20.0"
loguru = "^0.6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.971"
black = "^22.6.0"
isort = "^5.10.1"
flake8 = "^5.0.4"

[tool.poetry.scripts]
valence_api = "valence_api:cli.app"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"